Moved the EPS_FREECAD constant to geometry.constants; Identified and fixed the bug that caused the Interpolator's geometry to be considered infeasible (accidentally set the threshold too tight when the atol=0 overrode the default tolerance.).
